KOSPI_KOSDAQ_Analysis Project
Environment : Jupyter Notebook(Anaconda 3)
version 1.0
* You should change the font setting fitting with your font-having state.

This project is to analyze companies KOSPI and KOSDAQ  market cap ranked within top 50. It was scheduled to crawl all the data of 60 days, but I upload the 20 days version because the time to crawl all the data of 60 days is too long.
Just run the program in Jupyter Notebook, and you can see the flow of aspects of Korea stock market.